chore: next batch of the review fixes

+++ contracts/interfaces/IBuybackAllocator.sol
@@ -10,12 +10,19 @@ import {IOracleRouter} from "./IOracleRouter.sol";
  * @notice Public surface of the BuybackAllocator.
  */
 interface IBuybackAllocator {
+    /// @notice Whether a release can proceed, or why it is skipped. Reported in the skip event.
     enum AllocationStatus {
+        // The release can proceed
         Eligible,
+        // No budget available to spend
         NoAvailableBudget,
+        // The oracle returned no price
         QuoteUnavailable,
+        // The price is below the floor
         StEthPriceBelowMin,
+        // The spendable amount is below the smallest allowed
         AllocationBelowMin,
+        // The daily or yearly cap leaves no room
         WindowCapReached
     }
 
@@ -64,17 +71,17 @@ interface IBuybackAllocator {
 
     function allocate() external;
 
-    function setSurplusShareBP(uint16 surplusShareBP_) external;
+    function setSurplusShareBP(uint256 surplusShareBP_) external;
 
-    function setReserveDailyRateUSD(uint128 reserveDailyRateUSD_) external;
+    function setReserveDailyRateUSD(uint256 reserveDailyRateUSD_) external;
 
-    function setDailyCapUSD(uint128 dailyCapUSD_) external;
+    function setDailyCapUSD(uint256 dailyCapUSD_) external;
 
-    function setYearlyCapUSD(uint128 yearlyCapUSD_) external;
+    function setYearlyCapUSD(uint256 yearlyCapUSD_) external;
 
-    function setMinStEthPriceUSD(uint128 minStEthPriceUSD_) external;
+    function setMinStEthPriceUSD(uint256 minStEthPriceUSD_) external;
 
-    function setMinSpendPerCallUSD(uint128 minSpendPerCallUSD_) external;
+    function setMinSpendPerCallUSD(uint256 minSpendPerCallUSD_) external;
 
     function setExecutor(address newExecutor_) external;
